Stanford, Code in Class:  Final Project
This game is a tribute to my childhood love of Atari’s tennis and Pong games from the 1980s and is the final project for Standford's Code in Class. I set out to recreate that simple yet addictive experience on modern laptops and desktops, using the up and down arrow keys to control the paddle and challenging an AI opponent to see who reaches seven points first. Starting this class with no coding background, I spent nearly 30 hours writing the Python code in PyCharm—facing plenty of frustration along the way but determined to see it through. Because the Stanford IDE doesn’t support all the packages I needed, I needed a standalone file that lives outside any development environment. Usually, playing a Python program requires installing Python on your computer, but I wanted to make it as easy as possible for anyone with a Windows PC. To achieve that, I bundled all the necessary code into an application file: double-click “Bangee Pong Final.exe” to launch the game, or open “Bangee Pong.py” if you’d like to inspect the code.